refactor: decomposition
