+++ vite-project/src/pages/homepage.tsx
@@ -1,53 +1,13 @@
 import axios, { AxiosError, AxiosResponse } from 'axios';
-import React, { ChangeEvent, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
+import Modal from '../components/Modal';
+import Search from '../components/Search';
 import SearchCard from '../components/SearchCard';
+import { ICharacter, IResult, IEpisode } from '../servises/types';
 import '../styles/Modal.css';
 import '../styles/Search.css';
 
-interface ICharacter {
-  id: number;
-  name: string;
-  status: string;
-  species: string;
-  type: string;
-  gender: string;
-  origin: {
-    name: string;
-    url: string;
-  };
-  location: {
-    name: string;
-    url: string;
-  };
-  image: string;
-  episode: string[];
-  url: string;
-  created: string;
-}
-
-interface IResult {
-  info: {
-    count: number;
-    pages: number;
-    next: string | null;
-    prev: string | null;
-  };
-  results: ICharacter[];
-}
-
-interface IEpisode {
-  id: number;
-  name: string;
-  air_date: string;
-  episode: string;
-  characters: string[];
-  url: string;
-  created: string;
-}
-
 const Homepage = () => {
-  const [searchValue, setSearchValue] = useState<string>('');
-  const [isLoading, setIsLoading] = useState<boolean>(false);
   const [results, setResults] = useState<ICharacter[]>([]);
   const [isModal, setIsModal] = useState<boolean>(false);
   const [modalCharacter, setModalCharacter] = useState<ICharacter>(results[0]);
@@ -63,35 +23,20 @@ const Homepage = () => {
     fetchData().catch(console.error);
   }, []);
 
-  const handleSubmit = async (e: ChangeEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setIsLoading(true);
-    try {
-      const response: AxiosResponse<IResult> = await axios.get(
-        `https://rickandmortyapi.com/api/character/?name=${searchValue}`
-      );
-      setResults(response.data.results);
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        if (error.response?.status === 404) {
-          setResults([]);
-        } else console.error(error);
-      }
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { value } = e.target;
-    setSearchValue(value);
-  };
-
   const openModal = (id: number) => {
     setIsModal(true);
     const character = results.filter((i) => i.id === id);
     setModalCharacter(character[0]);
     getEpisodes(character[0]);
+  };
+
+  const closeModal = () => {
+    setIsModal(false);
+  };
+
+  const search = (results: ICharacter[]) => {
+    console.log('work' + results);
+    setResults(results);
   };
 
   const getEpisodes = (character: ICharacter) => {
@@ -113,23 +58,8 @@ const Homepage = () => {
   return (
     <div className="home_container">
       <h1>RS School React</h1>
+      <Search search={search} />
 
-      <div className="page-wrap">
-        <form className="search_field" onSubmit={handleSubmit}>
-          <label htmlFor="search">
-            <input
-              id="search"
-              type="text"
-              value={searchValue}
-              onChange={handleChange}
-              className="searchInput"
-            />
-          </label>
-          <button className="search_btn" type="submit" disabled={isLoading}>
-            {isLoading ? 'Loading...' : 'Search'}
-          </button>
-        </form>
-      </div>
       <div className="cards_container">
         {results.length !== 0 ? (
           results?.map((card, i) => (
@@ -150,48 +80,22 @@ const Homepage = () => {
           <h2>There is nothing here</h2>
         )}
       </div>
-      {isModal ? (
-        <div
-          className="modal_container"
-          onClick={(e) => e.currentTarget === e.target && setIsModal(false)}
-        >
-          <div className="modal_wrapper">
-            <div className="modal_window">
-              <div className="close" onClick={() => setIsModal(false)}>
-                ✖
-              </div>
-              <div className="modal_info">
-                <div className="modal_main_info">
-                  <h2>{modalCharacter.name}</h2>
-                  <div className="modal_picture">
-                    <img className="modal_img" src={modalCharacter.image} />
-                  </div>
-                </div>
-                <div className="modal_additional_info">
-                  <h5 className="card_subheader">Species: {modalCharacter.species}</h5>
-                  <div className="card_text">Gender: {modalCharacter.gender}</div>
-                  {modalCharacter.type ? (
-                    <div className="card_text">Type: {modalCharacter.type}</div>
-                  ) : (
-                    <div className="card_text">No type</div>
-                  )}
-                  <div className="card_text">Status: {modalCharacter.status}</div>
-                  <div className="card_text">Location: {modalCharacter.location.name}</div>
-                  <div className="card_text">Origin: {modalCharacter.origin.name}</div>
-                  <div className="card_text">Episode(s): {modalEpisodes}</div>
-                  <div className="card_text">url: {modalCharacter.url}</div>
-                  <div className="card_text">created: {modalCharacter.created}</div>
-                </div>
-              </div>
-              <button className="card_btn" onClick={() => setIsModal(false)}>
-                close
-              </button>
-            </div>
-          </div>
-        </div>
+      {isModal && (
+        <Modal
+          modalCharacter={modalCharacter}
+          modalEpisodes={modalEpisodes}
+          closeModal={closeModal}
+        />
+      )}
+      {/* {isModal ? (
+        <Modal
+          modalCharacter={modalCharacter}
+          modalEpisodes={modalEpisodes}
+          closeModal={closeModal}
+        />
       ) : (
         <div></div>
-      )}
+      )} */}
     </div>
   );
 };
