--- vite-project/src/servises/axios.ts
@@ -1,8 +1,0 @@
-import axios from 'axios';
-
-const axiosInstance = axios.create({
-  baseURL: 'https://rickandmortyapi.com/api',
-  timeout: 5000,
-});
-
-export default axiosInstance;
